scheduler/pyproject.toml
[project]
name = "scheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.11.3",
    "pyyaml>=6.0.2",
]
